Restructure validation interfaces and change validation utility from an abstract class to a trait.
