--- includes/client/cli/command/validate/Validator.php
@@ -1,16 +1,0 @@
-<?php
-namespace Press_Sync\client\cli\command\validate;
-
-/**
- * Interface Validator
- *
- * @package Press_Sync\client\cli\command\validate
- * @since NEXT
- */
-interface Validator {
-	/**
-	 * @return void
-	 * @since NEXT
-	 */
-	public function validate();
-}
