+++ includes/validation/ValidationUtility.php
@@ -10,15 +10,17 @@ namespace Press_Sync\validators;
 /**
  * This class defines utility methods that are common to all Validators.
  */
-abstract class Validation_Utility extends Validation_API {
+trait ValidationUtility {
 	/**
 	 * Get an icon based on whether a result was (bool) true or not.
 	 *
 	 * @since NEXT
-	 * @param  bool   $result The result to test.
+	 *
+	 * @param  bool $result The result to test.
+	 *
 	 * @return string
 	 */
 	protected function get_result_icon( $result ) {
-		return ((bool) $result) === true ? '✅' : '❌';
+		return ( (bool) $result ) === true ? '✅' : '❌';
 	}
 }
